Removes all skipped notes
