+++ src/output_queue/output_queue.py
@@ -3,11 +3,11 @@ import mido
 import time
 import platform
 from threading import Thread
-from multiprocessing import Process, Manager
+from multiprocessing import Lock, Process, Manager
 
 from src.common.midi_event import MidiEvent
 from src.common.shared_priority_queue import PeekingPriorityQueue
-from src.communication.messages import Message, MessageType, NoteOutputMessage, PlayingState
+from src.communication.messages import Message, MessageType, NoteOutputMessage, PlayingState, TimeSkipMessageType
 from src.output_queue.output_comm import OutputCommSystem
 from src.output_queue.synth import MIDISynthesizer, SYNTHESIZER_NAME
 from src.output_queue.tempo_mode import TempoMode
@@ -18,6 +18,7 @@ WINDOWS_SYNTH_KEYWORDS = "microsoft gs wavetable synth"
 class OutputQueue():
     def __init__(self, input_queue, output_queue):
         self.queue = PeekingPriorityQueue()
+        self.played_queue = PeekingPriorityQueue()
         self._open_port = None
         self.active = False
         self.last_note_timestamp = 0
@@ -28,12 +29,14 @@ class OutputQueue():
         self.state = PlayingState.STOP
         self.state_changed = False
         self.playing_notes = {}
+        self.accessLock = Lock()
 
         self.comm_system = OutputCommSystem()
         self.comm_system.set_queues(input_queue, output_queue)
         self.comm_system.registerListener(MessageType.SYSTEM_STOP, self.deactivate)
         self.comm_system.registerListener(MessageType.OUTPUT_QUEUE_UPDATE, self.process_note_event)
         self.comm_system.registerListener(MessageType.STATE_UPDATE, self.stateChanged)
+        self.comm_system.registerListener(MessageType.TIME_SKIP, self.timeSkip)
         self.comm_system.start()
 
         # TODO DJA-PROC Add a mode system to allow for different play modes
@@ -49,6 +52,25 @@ class OutputQueue():
             self.state_changed = True
 
         self.state = message.data
+
+    def timeSkip(self, message : Message):
+        if message.data == TimeSkipMessageType.FORWARD:
+            self.skip_forward(10)
+        elif message.data == TimeSkipMessageType.BACKWARD:
+            print("SKIP BACKWARD")
+
+    def skip_forward(self, seconds : int):
+        try:
+            with self.accessLock:
+                current_time = self.queue.peek().timestamp
+                print("Current Time: " + str(current_time))
+                while self.queue.peek().timestamp < current_time + seconds:
+                        print("Skipping: " + str(self.queue.peek().timestamp))
+                        self.played_queue.put(self.queue.get_nowait())
+        except queue.Empty as e:
+            pass # Expected if we dont have anything in the queue
+        except IndexError as e:
+            pass # Expected if we dont have anything in the queue
 
     # Selects the output device to send MIDI to. If `name` is None then the system default is used
     def select_device(self, name):
@@ -92,64 +114,66 @@ class OutputQueue():
 
     # Checks the queue for messages and sends them to the output as needed and returns the number of message sent (mainly for testing)
     def _check_priority_queue(self):
-        if self._open_port == None:
-            return
+        with self.accessLock:
+            if self._open_port == None:
+                return
 
-        now = time.time()
+            now = time.time()
 
-        # How many seconds into the song we are
-        relative_time = now - self.last_note_time_played + self.last_note_timestamp
+            # How many seconds into the song we are
+            relative_time = now - self.last_note_time_played + self.last_note_timestamp
 
-        immediate_events = []
-        button_events = []
+            immediate_events = []
+            button_events = []
 
-        # Separate events from buttons vs events from the song
-        while self.queue.qsize() > 0:
-            event = self.queue.peek()
+            # Separate events from buttons vs events from the song
+            while self.queue.qsize() > 0:
+                event = self.queue.peek()
 
-            if event.from_user_input:
-                button_events += [self.queue.get()]
-            elif self.state == PlayingState.PLAY and event.timestamp <= relative_time:
-                immediate_events += [self.queue.get()]
-            else:
-                break
+                if event.from_user_input:
+                    button_events += [self.queue.get()]
+                elif self.state == PlayingState.PLAY and event.timestamp <= relative_time:
+                    immediate_events += [self.queue.get()]
+                else:
+                    break
 
-        if self.state_changed:
-            self.state_changed = False
+            if self.state_changed:
+                self.state_changed = False
 
-            if self.state == PlayingState.PLAY:
-                self.play()
-            elif self.state == PlayingState.PAUSE:
-                self.pause()
-            elif self.state == PlayingState.STOP:
-                self.stop()
+                if self.state == PlayingState.PLAY:
+                    self.play()
+                elif self.state == PlayingState.PAUSE:
+                    self.pause()
+                elif self.state == PlayingState.STOP:
+                    self.stop()
 
-        if self.playing_mode is not None:
-            self.playing_mode.update(immediate_events, button_events, relative_time)
+            if self.playing_mode is not None:
+                self.playing_mode.update(immediate_events, button_events, relative_time)
 
-        for button_event in button_events:
-            self._send_midi_event(button_event)
+            for button_event in button_events:
+                self._send_midi_event(button_event)
 
-        # Handle regular queue events
-        for midiEvent in immediate_events:
-            # A timestamp of 0 means the note should be played immediately and is
-            # valid, but would otherwise mess up the timings
-            if midiEvent.timestamp > 1e-4:
-                self.last_note_time_played = now
-                self.last_note_timestamp = midiEvent.timestamp
+            # Handle regular queue events
+            for midiEvent in immediate_events:
+                # A timestamp of 0 means the note should be played immediately and is
+                # valid, but would otherwise mess up the timings
+                if midiEvent.timestamp > 1e-4:
+                    self.last_note_time_played = now
+                    self.last_note_timestamp = midiEvent.timestamp
 
-            if midiEvent.event.type == "note_off":
-                if midiEvent.event.note in self.playing_notes.keys():
-                    del self.playing_notes[midiEvent.event.note]
-            elif midiEvent.event.type == "note_on" and midiEvent.play_note:
-                self.playing_notes[midiEvent.event.note] = midiEvent.event
+                if midiEvent.event.type == "note_off":
+                    if midiEvent.event.note in self.playing_notes.keys():
+                        del self.playing_notes[midiEvent.event.note]
+                elif midiEvent.event.type == "note_on" and midiEvent.play_note:
+                    self.playing_notes[midiEvent.event.note] = midiEvent.event
 
-            if midiEvent.event.type == "note_on" and midiEvent.play_note:
-                self._send_midi_event(midiEvent)
-            elif midiEvent.event.type == "note_on" and not midiEvent.play_note:
-                self.comm_system.send(Message(MessageType.NOTE_OUTPUT, NoteOutputMessage(midiEvent, relative_time, now)))
-            else:
-                self._send_midi_event(midiEvent)
+                if midiEvent.event.type == "note_on" and midiEvent.play_note:
+                    self._send_midi_event(midiEvent)
+                elif midiEvent.event.type == "note_on" and not midiEvent.play_note:
+                    self.comm_system.send(Message(MessageType.NOTE_OUTPUT, NoteOutputMessage(midiEvent, relative_time, now)))
+                    self.played_queue.put(midiEvent)
+                else:
+                    self._send_midi_event(midiEvent)
 
     def _send_midi_event(self, midiEvent: MidiEvent):
         if type(midiEvent) != MidiEvent:
@@ -160,6 +184,7 @@ class OutputQueue():
 
         self._open_port.send(midiEvent.event)
         self.comm_system.send(Message(MessageType.NOTE_OUTPUT, NoteOutputMessage(midiEvent, relative_now, now)))
+        self.played_queue.put(midiEvent)
 
     def play(self):
         if self.last_note_timestamp is not None:
